chore: constant outputPath moved to apiRoute
