+++ packages/preset-umi/src/features/apiRoute/apiRoute.ts
@@ -3,7 +3,8 @@ import { logger, winPath } from '@umijs/utils';
 import fs from 'fs';
 import { basename, join, resolve } from 'path';
 import { matchApiRoute } from './utils';
-import { TEMPLATES_DIR, OUTPUT_PATH } from '../../constants';
+import { TEMPLATES_DIR } from '../../constants';
+import { OUTPUT_PATH } from './constants';
 import type { IApi, IApiMiddleware } from '../../types';
 import DevServerAdapterBuild from './dev-server/esbuild';
 import VercelAdapterBuild from './vercel/esbuild';
