+++ packages/preset-umi/src/features/apiRoute/constants.ts
@@ -1,0 +1,1 @@
+export const OUTPUT_PATH = '.output/server/pages/api';
